Lots of changes! Biggest one is a huge benefit. We are reading Sznajder game reports for 2020-2021 (few zone exit matches in Sznajder's data here) and 2021-2022 (lots of matches... somewhere on the order of 270 games I think).
One new problem detected: we are reading very few zone exits successfully. Key problem here is that, for whatever reason, the file properties of all of the zone exits files are when I downloaded them from Corey Sznajder, not when Sznajader created the files. This ruins the logic in Join_Entries.R and results int the zone exits not joining into the play-by-play data correctly. This is going to be a hard problem to fix. Will work on this soon!